requirements.txt
appnope==0.1.2
backcall==0.2.0
certifi==2021.5.30
click==8.0.1
dataclasses==0.8
decorator==5.1.0
Flask==2.0.1
importlib-metadata==4.8.1
ipdb==0.13.9
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
python-dotenv==0.19.0
six==1.16.0
toml==0.10.2
traitlets==4.3.3
typing-extensions==3.10.0.2
wcwidth==0.2.5
Werkzeug==2.0.1
zipp==3.5.0
